+++ CHANGELOG.md
@@ -44,10 +44,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - - Added Warrior's Gift to Power Effect Builder.
 - - Added Empathy to Power Effect Builder.
 - - Added Elemental Manipulation to Power Effect Builder.
+- - Added Blind to Power Effect Builder.
 </details>  
 
 - - Enabled near-infinite duration on powers if no power points setting rule active.
-- - Add an AE to the caster if he is not the target, that gives the casters spellcasting trait a -1 for maintaining the power if the power is found and no power points is active.
+- - Add an AE to the caster if he is not the target, that gives the casters spellcasting trait a -1 for maintaining the power if the power is found and no power points is active. In case no power points is not used, an AE is added to the caster (if the power is found) thats sole purpose is to track and cancel maintained powers.
 - If a power effect is set up using the power effect builder, the caster gains the ability to delete the AE on all actors he previously added it to. That means if the caster drops the power (or the duration runs out) it will be dropped on each actor, giving full control over maintenance to the player.
 - Localisation for:
 - - Fear Table function.